+++ web/src/views/types.ts
@@ -1,0 +1,7 @@
+export interface NewSessionSettings {
+  statusCode?: number
+  responseDelay?: number
+  contentType?: string
+  responseContent?: Uint8Array
+  destroyCurrentSession?: boolean
+}