+++ web/webpack/webpack.dev.js
@@ -1,0 +1,12 @@
+/* global module require */
+
+const { merge } = require('webpack-merge')
+const common = require('./webpack.common.js')
+
+module.exports = merge(common, {
+  devtool: 'inline-source-map',
+  mode: 'development',
+  optimization: {
+    minimize: false,
+  }
+})