+++ web/webpack/webpack.prod.js
@@ -1,0 +1,11 @@
+/* global module require */
+
+const { merge } = require('webpack-merge')
+const common = require('./webpack.common.js')
+
+module.exports = merge(common, {
+  mode: 'production',
+  optimization: {
+    minimize: true,
+  },
+})