+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][keepachangelog] and this project adheres to [Semantic Versioning][semver].
 
+## UNRELEASED
+
+### Changed
+
+- Frontend building using WebPack
+- VueJS updated from v2 to v3
+- Bootstrap updated from v4 to v5
+- All frontend dependencies are now built-in (no external network requests are needed anymore)
+
+### Removed
+
+- `--public` flag (and env variable `PUBLIC_DIR`) support for `serve` sub-command
+
+### Fixed
+
+- A lot of small frontend issues
+
 ## v0.4.3
 
 ### Changed